backend-azure/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azurefunctions-extensions-http-fastapi
openai
azure-cosmos 
langchain-openai 
langchain-community
langchain-unstructured
unstructured[pdf,docx]